pyproject.toml
[project]
name = "tetriscli"
version = "0.1.0"
description = "Simple cli tetris utility"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "windows-curses>=2.4.1",
]
